requirements.txt
pyyaml
envparse